[CS-4543]: Update biometricSwith to be reusable
