+++ cardstack/src/components/BiometricSwitch/BiometricSwitch.tsx
@@ -1,13 +1,11 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { Switch } from 'react-native';
 
 import { Container, Text, Icon } from '@cardstack/components';
-import { colors } from '@cardstack/theme';
 import {
   colorStyleVariants,
   ThemeVariant,
 } from '@cardstack/theme/colorStyleVariants';
-import { Device } from '@cardstack/utils';
 
 import { strings } from './strings';
 import { useBiometricSwitch } from './useBiometricSwitch';
@@ -16,11 +14,6 @@ interface BiometricSwitchProps {
   variant: ThemeVariant;
 }
 
-const androidSwitchLightConfig = {
-  false: colors.backgroundLightGray,
-  true: '', // empty to use default value
-};
-
 export const BiometricSwitch = ({ variant }: BiometricSwitchProps) => {
   const {
     biometryIconProps,
@@ -28,15 +21,8 @@ export const BiometricSwitch = ({ variant }: BiometricSwitchProps) => {
     isBiometryEnabled,
     biometryAvailable,
     toggleBiometrySwitch,
-  } = useBiometricSwitch();
-
-  const trackColor = useMemo(
-    () =>
-      Device.isAndroid && variant === 'light'
-        ? androidSwitchLightConfig
-        : undefined,
-    [variant]
-  );
+    trackColor,
+  } = useBiometricSwitch(variant);
 
   // iconProps is undefined until biometry is ready.
   if (!biometryAvailable || !biometryIconProps) return null;
