+++ cardstack/src/components/BiometricSwitch/useBiometricSwitch.tsx
@@ -1,4 +1,4 @@
-import { useCallback } from 'react';
+import { useCallback, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { useBiometry } from '@cardstack/hooks/useBiometry';
@@ -6,8 +6,16 @@ import {
   toggleBiometry,
   selectBiometryEnabled,
 } from '@cardstack/redux/biometryToggleSlice';
+import { colors } from '@cardstack/theme';
+import { ThemeVariant } from '@cardstack/theme/colorStyleVariants';
+import { Device } from '@cardstack/utils';
 
-export const useBiometricSwitch = () => {
+const androidSwitchLightConfig = {
+  false: colors.backgroundLightGray,
+  true: '', // empty to use default value
+};
+
+export const useBiometricSwitch = (variant?: ThemeVariant) => {
   const { biometryAvailable, biometryLabel, biometryIconProps } = useBiometry();
 
   const dispatch = useDispatch();
@@ -17,7 +25,16 @@ export const useBiometricSwitch = () => {
     dispatch(toggleBiometry());
   }, [dispatch]);
 
+  const trackColor = useMemo(
+    () =>
+      Device.isAndroid && variant === 'light'
+        ? androidSwitchLightConfig
+        : undefined,
+    [variant]
+  );
+
   return {
+    trackColor,
     biometryLabel,
     biometryIconProps,
     isBiometryEnabled,
